john-wiegley-captions
This is now my finished version. There are a few things I would like you
to look at apart from already mentioned.

I realise that it is no use me drawing your attention to line numbers,
so I have started noting the times instead.

At 33:10:38 - should this be ghc or Ghc because it is at the beginning
of a sentence.  Also, I presume it's in lower carriage because its part
of a 'formula'.
Same question re 33.40.93