s/[inaud]/inaudible
Quoting this from the wiki:

With recorded speech it's sometimes hard to tell what the speaker is
saying. Our transcriptionist uses the word `inaudible` as a placeholder
for something that she couldn't make out. If you think that you can tell
what the speaker is saying, please replace `inaudible` with what you hear.
If you can't tell what the speaker is saying, please change `inaudible` to
`[inaudible]`. If the speaker is actually saying the word 'inaudible',
then just leave it unchanged!

https://github.com/fromPeerToPeer/captions/wiki/Proofreading-captions-as-a-technical-reviewer